algorithm rebuild, wake up and keep going!
